(refactor): Extract project folder actions to a custom hook in `ProjectsList`
